Forcing NAXIS keywords to be integers when building a wcs header

+++ pnicer/utils.py
@@ -518,8 +518,8 @@ def data2header(lon, lat, frame="icrs", proj_code="CAR", pixsize=1/3600, enlarge
 
     # Determine extent of data for this projection
     x, y = wcs.WCS(header).wcs_world2pix(lon, lat, 1)
-    naxis1 = np.ceil((x.max()) - np.floor(x.min())) * enlarge
-    naxis2 = np.ceil((y.max()) - np.floor(y.min())) * enlarge
+    naxis1 = (np.ceil((x.max()) - np.floor(x.min())) * enlarge).astype(int)
+    naxis2 = (np.ceil((y.max()) - np.floor(y.min())) * enlarge).astype(int)
 
     # Calculate pixel shift relative to centroid (caused by unisotropic distribution of sources)
     xdelta = (x.min() + x.max()) / 2
